requirements.txt
certifi==2016.2.28
Django==1.11.3
djangorestframework==3.6.4
dnspython==1.15.0
email-validator==1.0.2
Faker==0.8.3
idna==2.6
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
wincertstore==0.2
